+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.3.1
+
+#### :bug: Bug Fix
+
+* Fixed invalid imports
+
 ## v3.3.0
 
 #### :rocket: New Feature